requirements.txt
asgiref==3.4.1
autopep8==1.5.7
Babel==2.9.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
cryptography==3.4.8
defusedxml==0.7.1
dj-rest-auth==2.1.11
Django==3.2.7
django-allauth==0.45.0
django-countries==7.2.1
django-environ==0.7.0
django-money==2.1
django-phonenumber-field==5.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
idna==3.2
oauthlib==3.1.1
phonenumbers==8.12.31
psycopg2==2.9.1
py-moneyed==1.2
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.6
